third_party/WebKit/LayoutTests/http/tests/devtools/editor/text-editor-smart-braces-expected.txt

This test checks text editor smart braces functionality.


Running: testTypeBraceSequence
({[|]})

Running: testBraceOverride
({[]})|

Running: testQuotesToCloseStringLiterals
'Hello"'|

Running: testQuotesToCloseStringLiteralInsideLine
console.log("information"|);
